+++ src/Lake/Diagnostics/LakeBuildLog.cs
@@ -30,14 +30,16 @@ namespace Lake.Diagnostics
         /// </summary>
         /// <param name="verbosity">The verbosity.</param>
         /// <param name="level">The level.</param>
-        /// <param name="message">The message.</param>
-        public void Write(Verbosity verbosity, LogLevel level, string message)
+        /// <param name="format">A composite format string.</param>
+        /// <param name="args">An object array that contains zero or more objects to format.</param>
+        public void Write(Verbosity verbosity, LogLevel level, string format, params object[] args)
         {
             if (verbosity > Verbosity)
             {
                 return;
-            }            
-            _log.Write(verbosity, level, string.Format("[{0}] {1}", level.ToString().Substring(0, 1), message));
+            }
+            format = string.Concat("[", level.ToString().Substring(0, 1), "] ", format);
+            _log.Write(verbosity, level, format, args);
         }
     }
 }