+++ src/Lunt/Diagnostics/ConsoleBuildLog.cs
@@ -23,13 +23,14 @@ namespace Lunt.Diagnostics
         /// </summary>
         /// <param name="verbosity">The verbosity.</param>
         /// <param name="level">The level.</param>
-        /// <param name="message">The message.</param>
-        public void Write(Verbosity verbosity, LogLevel level, string message)
+        /// <param name="format">A composite format string.</param>
+        /// <param name="args">An object array that contains zero or more objects to format.</param>
+        public void Write(Verbosity verbosity, LogLevel level, string format, params object[] args)
         {
             try
             {
                 _writer.SetForeground(GetColor(level));
-                _writer.WriteLine(message);
+                _writer.WriteLine(format, args);
             }
             finally
             {